rename naics.py as sectors.py
