+++ flowsa/data_source_scripts/BLS_QCEW.py
@@ -20,7 +20,7 @@ from flowsa.location import US_FIPS
 from flowsa.flowbyfunctions import assign_fips_location_system
 from flowsa.flowbyactivity import FlowByActivity
 from flowsa.flowsa_log import log
-from flowsa.naics import industry_spec_key
+from flowsa.sectors import industry_spec_key
 
 
 def BLS_QCEW_URL_helper(*, build_url, year, **_):
