+++ flowsa/data_source_scripts/USDA_ERS_MLU.py
@@ -23,7 +23,7 @@ from flowsa.literature_values import \
     get_urban_land_use_for_railroads, get_open_space_fraction_of_urban_area
 from flowsa.validation import compare_df_units
 from flowsa.flowbyactivity import FlowByActivity
-from flowsa.naics import industry_spec_key
+from flowsa.sectors import industry_spec_key
 
 
 def mlu_call(*, resp, **_):
